tag rules

+++ CustomerTags/Controller/Adminhtml/Tags/MassDelete.php
@@ -7,6 +7,8 @@ use Magento\Backend\App\Action\Context;
 use Magento\Ui\Component\MassAction\Filter;
 use RookieSoft\CustomerTags\Model\ResourceModel\Tag\CollectionFactory;
 use Magento\Framework\App\ResponseInterface;
+use RookieSoft\CustomerTags\Model\ResourceModel\EntityTag\CollectionFactory as EntityTagCollectionFactory;
+use RookieSoft\CustomerTags\Model\TagRuleFactory as TagRuleFactory;
 
 class MassDelete extends \Magento\Backend\App\Action
 {
@@ -29,11 +31,15 @@ class MassDelete extends \Magento\Backend\App\Action
     public function __construct(
         Context $context,
         Filter $filter,
-        CollectionFactory $collectionFactory
-    ) {
+        CollectionFactory $collectionFactory,
+        EntityTagCollectionFactory $entityTagCollectionFactory,
+        TagRuleFactory $tagRuleFactory
 
+    ) {
         $this->_filter = $filter;
         $this->_collectionFactory = $collectionFactory;
+        $this->_entityTagCollectionFactory = $entityTagCollectionFactory;
+        $this->_tagRuleFactory = $tagRuleFactory;
         parent::__construct($context);
     }
 
@@ -45,11 +51,18 @@ class MassDelete extends \Magento\Backend\App\Action
         $collection = $this->_filter->getCollection($this->_collectionFactory->create());
         $collectionSize = $collection->getSize();
         foreach ($collection as $item) {
+            $tagCodes = $this->_entityTagCollectionFactory->create()->addFieldToFilter('tag_Code', $item->getCode())->addFieldToFilter('entity_type_id', 4);
+
+            if($tagCodes != NULL){
+                foreach($tagCodes as $tagCode){
+                    $tagRule =  $this->_tagRuleFactory->create()->load($tagCode->getEntity_id());
+                    $this->messageManager->addErrorMessage(__('the tag: '.$item->getLabel().' is still used in the rule: '.$tagRule->getRuleName()));
+                }
+                return $this->resultFactory->create(ResultFactory::TYPE_REDIRECT)->setPath('*/*/index');
+            }
             $item->delete();
+            $this->messageManager->addSuccess(__('A total of %1 record(s) have been deleted.', $collectionSize));
         }
-
-        $this->messageManager->addSuccess(__('A total of %1 record(s) have been deleted.', $collectionSize));
-
         return $this->resultFactory->create(ResultFactory::TYPE_REDIRECT)->setPath('*/*/index');
     }
